Panics in state critical routes terminates the process
A route managing critical state required to operate the workload
correctly is called a state critical route.
If a panic occurs while handling a request in a state critical route,
the process is terminated.
Panics are triggered if an assertion of an invariant is failed.
For example, when managing workers, workers, heartbeats and
shard count data structured should be kept in sync.
When this is not the case, it's not safe to continue to operate.